Sépare les fonctions d'affichage des notifications "toast"
